Moved nodes around
